+++ src/magnetic_interaction/mod.rs
@@ -9,7 +9,7 @@ pub struct Force();
 /// Returns force on unit magnetic moment with orientation `o` in a given field
 /// gradient `grad_b`.
 ///
-pub fn mean_force(grad_b: ArrayView<f64, Ix2>, o: &OrientationVector) -> Vector<Force> {
+pub fn mean_force(grad_b: ArrayView<f32, Ix2>, o: &OrientationVector) -> Vector<Force> {
     let o = Array::from_vec(o.v.to_vec());
     let f = grad_b.dot(&o);
     unsafe{[
